+++ packages/composites/anchor/src/config.ts
@@ -4,17 +4,22 @@ import { Config as DefaultConfig } from '@chainlink/types'
 export const DEFAULT_ENDPOINT = 'price'
 export const DEFAULT_TOKEN_DECIMALS = 18
 
+export const DEFAULT_ANCHOR_VAULT_CONTRACT_ADDRESS = '0xA2F987A546D4CD1c607Ee8141276876C26b72Bdf'
+export const DEFAULT_TERRA_BLUNA_CONTRACT_ADDRESS = 'terra1mtwph2juhj0rvjz7dy92gvl6xvukaxu8rfv8ts'
+
 export interface Config extends DefaultConfig {
   anchorVaultContractAddress: string
+  terraBLunaContractAddress: string
 }
 
 export const makeConfig = (prefix?: string): Config => {
-  const config = Requester.getDefaultConfig(prefix)
-  config.rpcUrl = util.getRequiredEnv('RPC_URL', prefix)
-  config.defaultEndpoint = DEFAULT_ENDPOINT
-
   return {
-    ...config,
-    anchorVaultContractAddress: util.getRequiredEnv('ANCHOR_VAULT_CONTRACT_ADDRESS', prefix),
+    ...Requester.getDefaultConfig(prefix),
+    rpcUrl: util.getRequiredEnv('RPC_URL', prefix),
+    defaultEndpoint: util.getEnv('API_ENDPOINT', prefix) || DEFAULT_ENDPOINT,
+    anchorVaultContractAddress:
+      util.getEnv('ANCHOR_VAULT_CONTRACT_ADDRESS', prefix) || DEFAULT_ANCHOR_VAULT_CONTRACT_ADDRESS,
+    terraBLunaContractAddress:
+      util.getEnv('TERRA_BLUNA_CONTRACT_ADDRESS', prefix) || DEFAULT_TERRA_BLUNA_CONTRACT_ADDRESS,
   }
 }