+++ packages/core/legos/src/sources.ts
@@ -62,6 +62,7 @@ import * as gemini from '@chainlink/gemini-adapter'
 import * as genesis_adapter from '@chainlink/genesis-volatility-adapter'
 import * as geodb from '@chainlink/geodb-adapter'
 import * as google_bigquery from '@chainlink/google-bigquery-adapter'
+import * as gramchain from '@chainlink/gramchain-adapter'
 import * as graphql from '@chainlink/graphql-adapter'
 import * as iex_adapter from '@chainlink/iex-cloud-adapter'
 import * as intrinio from '@chainlink/intrinio-adapter'
@@ -77,6 +78,7 @@ import * as lotus from '@chainlink/lotus-adapter'
 import * as marketstack from '@chainlink/marketstack-adapter'
 import * as messari from '@chainlink/messari-adapter'
 import * as metalsapi from '@chainlink/metalsapi-adapter'
+import * as mock_ea from '@chainlink/mock-ea-adapter'
 import * as mycryptoapi from '@chainlink/mycryptoapi-adapter'
 import * as ncfx from '@chainlink/ncfx-adapter'
 import * as nikkei from '@chainlink/nikkei-adapter'
@@ -94,6 +96,7 @@ import * as renvm_adapter_address from '@chainlink/renvm-address-set-adapter'
 import * as satoshitango from '@chainlink/satoshitango-adapter'
 import * as snowflake from '@chainlink/snowflake-adapter'
 import * as sochain from '@chainlink/sochain-adapter'
+import * as solana_view_function from '@chainlink/solana-view-function-adapter'
 import * as spectral_macro_score from '@chainlink/spectral-macro-score-adapter'
 import * as sportsdataio from '@chainlink/sportsdataio-adapter'
 import * as stasis from '@chainlink/stasis-adapter'
@@ -182,6 +185,7 @@ export default {
   genesis_adapter,
   geodb,
   google_bigquery,
+  gramchain,
   graphql,
   iex_adapter,
   intrinio,
@@ -197,6 +201,7 @@ export default {
   marketstack,
   messari,
   metalsapi,
+  mock_ea,
   mycryptoapi,
   ncfx,
   nikkei,
@@ -214,6 +219,7 @@ export default {
   satoshitango,
   snowflake,
   sochain,
+  solana_view_function,
   spectral_macro_score,
   sportsdataio,
   stasis,