+++ packages/sources/gramchain/test/integration/adapter.test.ts
@@ -1,0 +1,72 @@
+import { AdapterRequest } from '@chainlink/types'
+import request, { SuperTest, Test } from 'supertest'
+import process from 'process'
+import nock from 'nock'
+import http from 'http'
+import { server as startServer } from '../../src'
+import { mockGramChainResponseSuccess } from './fixtures'
+import { DEFAULT_BASE_URL } from '../../src/config'
+import { AddressInfo } from 'net'
+
+let oldEnv: NodeJS.ProcessEnv
+
+beforeAll(() => {
+  oldEnv = JSON.parse(JSON.stringify(process.env))
+  process.env.CACHE_ENABLED = 'false'
+  process.env.API_ENDPOINT = process.env.API_ENDPOINT || DEFAULT_BASE_URL
+  process.env.API_VERBOSE = 'true'
+  if (process.env.RECORD) {
+    nock.recorder.rec()
+  }
+})
+
+afterAll(() => {
+  process.env = oldEnv
+  if (process.env.RECORD) {
+    nock.recorder.play()
+  }
+
+  nock.restore()
+  nock.cleanAll()
+  nock.enableNetConnect()
+})
+
+describe('execute', () => {
+  const id = '1'
+  let server: http.Server
+  let req: SuperTest<Test>
+
+  beforeAll(async () => {
+    server = await startServer()
+    req = request(`localhost:${(server.address() as AddressInfo).port}`)
+    process.env.CACHE_ENABLED = 'false'
+  })
+
+  afterAll((done) => {
+    server.close(done)
+  })
+
+  describe('api', () => {
+    const data: AdapterRequest = {
+      id,
+      data: {
+        custodianID: 'Cache',
+        metalCode: 'AU',
+        utilizationLockCode: 'Locked',
+      },
+    }
+
+    it('should return success', async () => {
+      mockGramChainResponseSuccess()
+
+      const response = await req
+        .post('/')
+        .send(data)
+        .set('Accept', '*/*')
+        .set('Content-Type', 'application/json')
+        .expect('Content-Type', /json/)
+        .expect(200)
+      expect(response.body).toMatchSnapshot()
+    })
+  })
+})