+++ packages/core/reference-data-reader/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @chainlink/ea-reference-data-reader
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [eecdac90b]
+  - @chainlink/ea-bootstrap@1.11.0
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [57be274ff]
+- Updated dependencies [540e563a9]
+- Updated dependencies [72f96124d]
+  - @chainlink/ea-bootstrap@1.10.6
+
 ## 1.1.1
 
 ### Patch Changes