+++ packages/sources/anyblock/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/anyblock-adapter
 
+## 1.1.13
+
+### Patch Changes
+
+- Updated dependencies [eecdac90b]
+  - @chainlink/ea-bootstrap@1.11.0
+  - @chainlink/ea-test-helpers@1.1.0
+
+## 1.1.12
+
+### Patch Changes
+
+- Updated dependencies [57be274ff]
+- Updated dependencies [540e563a9]
+- Updated dependencies [72f96124d]
+  - @chainlink/ea-bootstrap@1.10.6
+  - @chainlink/ea-test-helpers@1.1.0
+
 ## 1.1.11
 
 ### Patch Changes