+++ packages/sources/json-rpc/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @chainlink/json-rpc-adapter
 
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [eecdac90b]
+  - @chainlink/ea-bootstrap@1.11.0
+  - @chainlink/ea-test-helpers@1.1.0
+
+## 1.2.0
+
+### Minor Changes
+
+- 048a65b7b: update ea to newest pattern
+
+### Patch Changes
+
+- Updated dependencies [57be274ff]
+- Updated dependencies [540e563a9]
+- Updated dependencies [72f96124d]
+  - @chainlink/ea-bootstrap@1.10.6
+  - @chainlink/ea-test-helpers@1.1.0
+
 ## 1.1.22
 
 ### Patch Changes